requirements.txt
streamlit>=1.31.0
requests>=2.31.0
cryptography>=42.0.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
